requirements.txt
aiosmtplib==2.0.2
annotated-types==0.5.0
anyio==3.7.0
automata-lib==7.1.0
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
dnspython==2.4.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.100.0
fastapi-mail==1.4.0
frozendict==2.3.8
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
networkx==3.1
pika==1.3.2
prometheus-client==0.17.1
prometheus-fastapi-instrumentator==6.1.0
psycopg2==2.9.6
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
pydot==1.4.2
pyparsing==3.1.0
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
